Revert "Dagger dependensies refactoring"
This reverts commit c5fdcb878ae478bd6e95eda4c583fa231523ec94.
