+++ presentation/src/main/java/by/grodno/vasili/presentation/feature/notes/NotesActivityModule.java
@@ -2,6 +2,10 @@ package by.grodno.vasili.presentation.feature.notes;
 
 import android.arch.lifecycle.ViewModel;
 
+import by.grodno.vasili.data.datasource.FirebaseNoteEntityDatasource;
+import by.grodno.vasili.data.datasource.NoteEntityDatasource;
+import by.grodno.vasili.data.entity.mapper.NoteEntityDataMapper;
+import by.grodno.vasili.data.repository.NoteDataRepository;
 import by.grodno.vasili.domain.interactor.GetNotesListUseCase;
 import by.grodno.vasili.domain.repository.NoteRepository;
 import by.grodno.vasili.presentation.dagger.ViewModelKey;
@@ -17,7 +21,12 @@ public abstract class NotesActivityModule {
 
     @Provides
     @NotesActivityScope
-    static GetNotesListUseCase provideUserCase(IOThread ioThread, UIThread uiThread, NoteRepository repository) {
+    static GetNotesListUseCase provideUserCase() {
+        NoteEntityDataMapper mapper = new NoteEntityDataMapper();
+        NoteEntityDatasource datasource = new FirebaseNoteEntityDatasource(mapper);
+        NoteRepository repository = new NoteDataRepository(datasource, mapper);
+        UIThread uiThread = new UIThread();
+        IOThread ioThread = new IOThread();
         return new GetNotesListUseCase(ioThread, uiThread, repository);
     }
