+++ presentation/src/main/java/by/grodno/vasili/presentation/feature/notes/NotesAdapter.java
@@ -1,6 +1,5 @@
 package by.grodno.vasili.presentation.feature.notes;
 
-import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,8 +7,6 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import java.util.List;
-
-import javax.inject.Inject;
 
 import by.grodno.vasili.presentation.R;
 import by.grodno.vasili.presentation.model.NoteItem;
@@ -22,25 +19,20 @@ import static java.util.Collections.emptyList;
 class NotesAdapter extends RecyclerView.Adapter<NotesAdapter.ViewHolder> {
     private List<NoteItem> notes = emptyList();
 
-    @Inject
-    NotesAdapter() {
-    }
-
     public void setNotes(List<NoteItem> notes) {
         this.notes = notes;
         // TODO: Heavy operation
         notifyDataSetChanged();
     }
 
-    @NonNull
     @Override
-    public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+    public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         View itemView = LayoutInflater.from(parent.getContext()).inflate(R.layout.note_item, parent, false);
         return new ViewHolder(itemView);
     }
 
     @Override
-    public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
+    public void onBindViewHolder(ViewHolder holder, int position) {
         NoteItem item = notes.get(position);
         holder.idLabel.setText(item.id);
         holder.titleLabel.setText(item.title);
