+++ presentation/src/main/java/by/grodno/vasili/presentation/thread/IOThread.java
@@ -1,7 +1,5 @@
 package by.grodno.vasili.presentation.thread;
 
-import javax.inject.Inject;
-
 import by.grodno.vasili.domain.executor.SubscriberThread;
 import io.reactivex.Scheduler;
 import io.reactivex.schedulers.Schedulers;
@@ -10,10 +8,6 @@ import io.reactivex.schedulers.Schedulers;
  * Thread for IO-bound work
  */
 public class IOThread implements SubscriberThread {
-    @Inject
-    public IOThread() {
-    }
-
     @Override
     public Scheduler getScheduler() {
         return Schedulers.io();
