+++ presentation/src/main/java/by/grodno/vasili/presentation/thread/UIThread.java
@@ -1,7 +1,5 @@
 package by.grodno.vasili.presentation.thread;
 
-import javax.inject.Inject;
-
 import by.grodno.vasili.domain.executor.PostExecutionThread;
 import io.reactivex.Scheduler;
 import io.reactivex.android.schedulers.AndroidSchedulers;
@@ -10,10 +8,6 @@ import io.reactivex.android.schedulers.AndroidSchedulers;
  * Main thread for executing process for UI
  */
 public class UIThread implements PostExecutionThread {
-    @Inject
-    public UIThread() {
-    }
-
     @Override
     public Scheduler getScheduler() {
         return AndroidSchedulers.mainThread();
